Update list_countries.py
Added comments to "list_countries.py" under the get_countries(): to determine which countries have gotten hit very hard for covid-19. Which countries have no cases of covid-19 and which countries have a few cases. This will help the user who is trying to read the code to understand what the code is doing